Migrate x-pack-deprecation REST tests (#131444)
