+++ x-pack/plugin/deprecation/qa/deprecation-plugin/build.gradle
@@ -1,0 +1,20 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+apply plugin: 'elasticsearch.base-internal-es-plugin'
+apply plugin: 'elasticsearch.build'
+
+esplugin {
+  name = 'deprecation-plugin'
+  description = 'Deprecated query plugin'
+  classname ='org.elasticsearch.xpack.deprecation.plugin.TestDeprecationPlugin'
+}
+
+dependencies {
+  compileOnly project(":server")
+  implementation project(':x-pack:plugin:deprecation:qa::common')
+}
