--- x-pack/plugin/deprecation/qa/early-deprecation-rest/build.gradle
@@ -1,42 +1,0 @@
-import org.elasticsearch.gradle.util.GradleUtils
-import org.elasticsearch.gradle.internal.info.BuildParams
-
-apply plugin: 'elasticsearch.base-internal-es-plugin'
-apply plugin: 'elasticsearch.legacy-java-rest-test'
-
-esplugin {
-  description = 'Deprecated query plugin'
-  classname ='org.elasticsearch.xpack.deprecation.EarlyDeprecationTestPlugin'
-}
-
-dependencies {
-  javaRestTestImplementation project(path: ':x-pack:plugin:deprecation:qa:common')
-
-  javaRestTestImplementation("com.fasterxml.jackson.core:jackson-annotations:${versions.jackson}")
-  javaRestTestImplementation("com.fasterxml.jackson.core:jackson-databind:${versions.jackson}")
-}
-
-// let the javaRestTest see the classpath of main
-GradleUtils.extendSourceSet(project, "main", "javaRestTest", tasks.named("javaRestTest"))
-
-restResources {
-  restApi {
-    include '_common', 'indices', 'index'
-  }
-}
-
-testClusters.configureEach {
-  testDistribution = 'DEFAULT'
-  setting 'xpack.security.enabled', 'false'
-  setting 'xpack.license.self_generated.type', 'trial'
-  setting 'cluster.deprecation_indexing.enabled', 'true'
-  setting 'cluster.deprecation_indexing.flush_interval', '1ms'
-  setting 'logger.org.elasticsearch.xpack.deprecation','TRACE'
-  setting 'logger.org.elasticsearch.xpack.deprecation.logging','TRACE'
-}
-
-// Test clusters run with security disabled
-tasks.named("javaRestTest") {
-  buildParams.withFipsEnabledOnly(it)
-}
-
